Finish GetBlocks() api

+++ pkg/eth/eth.go
@@ -5,8 +5,8 @@ import (
 )
 
 type Eth interface {
-	GetBlocks(ctx context.Context, n int) ([]*Block, error)
-	GetBlock(ctx context.Context, id int) (*Block, error)
+	GetBlocks(ctx context.Context, n int64) ([]*Block, error)
+	GetBlock(ctx context.Context, id int64) (*Block, error)
 
 	GetTransation(ctx context.Context, txHash Hash) (*Transation, error)
 }
